Added Orb..
An Orb gives the Player a few extra jumps.
An Orb only adds to a Player if the player doesn't have that Orb yet or if the player has use at least 1 jump from that Orb.
An Orb keeps it's own counter for the amount of times the players used that orb.
Player implements Orb and uses it's Orb related variables to organize between Orbs.