Added Key.as, Door.as Textures.as..
Door.as is empty for now but it's functions are soon needed for Key.as.
Textures.as is only a singleton you call to load a texture.
Key.as is the first component which uses a texture.
Key.as does not have mechanics implemented yet.
Removed Player.as object from Main.as