Small adjustments to Key.as and Door.as
Door.as now only has a switchDoor function. This switches the door between collideable and non-collidable.
Key.as contructor has changed. It is no longer required to add doors as parameter.
addDoor and addDoors functions added in Key.as.
Key.as now calles switchDoor from Door.as.
switchDoor allowes doors to start open and close after hitting a key.