Adjustments to Finish.as and Key.as..
Key.as no sends a message wit MessageBoard.as when a key is reached.
Finish.as no requires a player when constructed.
Finish.as recieves messages from Key.as so it can keep currentKey up to date.
Finish.as uses a texture from Textures.as
Finish gives an error once you've won. The error has to be replaced by a win screen.